refactor(core): session routes
